+++ src/main/java/com/programmers/library/controller/MenuController.java
@@ -8,7 +8,7 @@ import java.util.function.Supplier;
 import com.programmers.library.exception.InvalidMenuException;
 import com.programmers.library.io.Input;
 import com.programmers.library.io.Output;
-import com.programmers.library.model.Menu;
+import com.programmers.library.enums.Menu;
 import com.programmers.library.service.LibarayManagerService;
 
 public class MenuController implements Runnable {
