+++ src/main/java/com/programmers/library/dto/BorrowBookRequest.java
@@ -1,11 +1,11 @@
-package com.programmers.library.model.request;
+package com.programmers.library.dto;
 
 import static com.programmers.library.constants.MessageConstants.*;
 
-public class ReturnBookRequest {
+public class BorrowBookRequest {
 	private long id;
 
-	public ReturnBookRequest(String id) {
+	public BorrowBookRequest(String id) {
 		try {
 			this.id = Long.parseLong(id);
 		} catch (NumberFormatException e) {
