+++ src/main/java/com/programmers/library/io/Input.java
@@ -1,13 +1,13 @@
 package com.programmers.library.io;
 
-import com.programmers.library.model.Menu;
-import com.programmers.library.model.Mode;
-import com.programmers.library.model.request.AddBookRequest;
-import com.programmers.library.model.request.BorrowBookRequest;
-import com.programmers.library.model.request.DeleteBookRequest;
-import com.programmers.library.model.request.LostBookRequest;
-import com.programmers.library.model.request.ReturnBookRequest;
-import com.programmers.library.model.request.FindBookRequest;
+import com.programmers.library.enums.Menu;
+import com.programmers.library.enums.Mode;
+import com.programmers.library.dto.AddBookRequest;
+import com.programmers.library.dto.BorrowBookRequest;
+import com.programmers.library.dto.DeleteBookRequest;
+import com.programmers.library.dto.LostBookRequest;
+import com.programmers.library.dto.ReturnBookRequest;
+import com.programmers.library.dto.FindBookRequest;
 
 public interface Input {
 	Mode inputMode();
