+++ src/main/java/com/programmers/library/service/LibarayManagerService.java
@@ -3,12 +3,12 @@ package com.programmers.library.service;
 import java.util.List;
 
 import com.programmers.library.entity.Book;
-import com.programmers.library.model.request.AddBookRequest;
-import com.programmers.library.model.request.BorrowBookRequest;
-import com.programmers.library.model.request.DeleteBookRequest;
-import com.programmers.library.model.request.LostBookRequest;
-import com.programmers.library.model.request.ReturnBookRequest;
-import com.programmers.library.model.request.FindBookRequest;
+import com.programmers.library.dto.AddBookRequest;
+import com.programmers.library.dto.BorrowBookRequest;
+import com.programmers.library.dto.DeleteBookRequest;
+import com.programmers.library.dto.LostBookRequest;
+import com.programmers.library.dto.ReturnBookRequest;
+import com.programmers.library.dto.FindBookRequest;
 
 public interface LibarayManagerService {
 	void addBook(AddBookRequest request);
