+++ src/main/java/com/programmers/library/util/FileUtils.java
@@ -1,11 +1,9 @@
 package com.programmers.library.util;
 
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
-import java.util.ArrayList;
 import java.util.List;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
