+++ src/test/java/com/programmers/library/dto/DeleteBookRequestTest.java
@@ -1,8 +1,10 @@
-package com.programmers.library.model.request;
+package com.programmers.library.dto;
 
 import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
+
+import com.programmers.library.dto.DeleteBookRequest;
 
 class DeleteBookRequestTest {
